Move auth routing to /auth namespace
