+++ api/actions/auth/index.js
@@ -1,0 +1,3 @@
+export login from './login';
+export logout from './logout'; 
+export signup from './signup';
