+++ api/actions/auth/logout.js
