+++ api/actions/index.js
@@ -1,6 +1,4 @@
 export loadInfo from './loadInfo';
 export loadAuth from './loadAuth';
-export signup from './signup';
-export login from './login';
-export logout from './logout';
 export * as widget from './widget/index';
+export * as auth from './auth/index';
